microsoft-edu-ai-hackathon-2025/backend/requirements.txt
flask
werkzeug
openai
requests
azure-identity
python-dotenv
pillow
numpy
pandas
opencv-python
ffmpeg-python
pypdf2
faster-whisper
torch
torchaudio
torchvision